requirements.txt
django
djangorestframework
pillow
djangorestframework-simplejwt
django-cors-headers